Fix the Array Component
Pages has an old array component that allows users to stuff
content for a pages key into an array.

The transition from form objects to FormData broke this functionality.
To fix this bug, the `FormPagesArray` component now encodes the
content_array as a string representation of a json object and the Teal
plugin decodes the content before saving it to the database.